requirements.txt
pyTelegramBotAPI>=4.21.0
PyYAML>=6.0.1
requests>=2.31.0
